requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.0
python-dotenv==1.2.2
sqlparse==0.5.5
